+++ station-container-api-v0/src/main/java/net/modificationstation/stationapi/impl/gui/screen/container/GuiHelperImpl.java
@@ -7,7 +7,7 @@ import net.modificationstation.stationapi.api.StationAPI;
 import net.modificationstation.stationapi.api.network.packet.MessagePacket;
 import net.modificationstation.stationapi.api.network.packet.PacketHelper;
 import net.modificationstation.stationapi.api.util.Identifier;
-import net.modificationstation.stationapi.impl.InventoryMessagePacket;
+import net.modificationstation.stationapi.impl.network.packet.play.InventoryMessagePacket;
 
 import java.util.function.Consumer;
